+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,8 +1,8 @@
-package roomescape.reservation.repository;
+package roomescape.repository;
 
 import java.util.List;
-import roomescape.reservation.controller.request.ReservationRequest;
-import roomescape.reservation.controller.response.ReservationResponse;
+import roomescape.controller.request.ReservationRequest;
+import roomescape.controller.response.ReservationResponse;
 
 public interface ReservationRepository {
